services/mobile-api-gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
grpcio==1.67.1
protobuf==5.28.3
